# project-management-ui/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi specification - Ziedzaa",
    "description": "OpenApi documentation for Spring Security",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "Ziedzaa",
      "url": "https://some-url.com",
      "email": "ziededucation@gmail.com"
    },
    "license": {
      "name": "Licence name",
      "url": "https://some-url.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:9090/api/v1",
      "description": "Local ENV"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/structures": {
      "post": {
        "tags": [
          "Structure"
        ],
        "operationId": "saveStructure",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StructureRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/projets": {
      "get": {
        "tags": [
          "Projet"
        ],
        "operationId": "findAllProjetsByUserStructure",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseProjetResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Projet"
        ],
        "operationId": "saveProjet",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
...
    "/structures/{structureId}": {
      "get": {
        "tags": [
          "Structure"
        ],
        "operationId": "findStructureById",
        "parameters": [
          {
            "name": "structureId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StructureResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/projets/{projet-id}": {
      "get": {
        "tags": [
          "Projet"
        ],
        "operationId": "findProjetById",
        "parameters": [
          {
            "name": "projet-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjetResponse"
                }
              }
            }
          }
        }
      }
    },
    "/projets/all": {
      "get": {
        "tags": [
          "Projet"
        ],
        "operationId": "findAllProjets",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseProjetResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/factures/projet/{projetId}": {
      "get": {
        "tags": [
          "Facture"
        ],
        "operationId": "getFacturesByProjetId",
        "parameters": [
          {
            "name": "projetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseFactureResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/contrats/projet/{projetId}": {
      "get": {
        "tags": [
          "Contrat"
        ],
        "operationId": "getContratsByProjetId",
        "parameters": [
          {
            "name": "projetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseContratResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/projets/{projetId}": {
      "delete": {
        "tags": [
          "Projet"
        ],
        "operationId": "deleteProjet",
        "parameters": [
          {
            "name": "projetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "StructureRequest": {
        "required": [
          "descriptionRequestDto",
          "titreRequestDto"
        ],
        "type": "object",
        "properties": {
          "idRequestDto": {
            "type": "integer",
            "format": "int32"
          },
          "referenceRequestDto": {
            "type": "string"
          },
          "titreRequestDto": {
            "maxLength": 30,
            "minLength": 2,
            "type": "string"
          },
          "descriptionRequestDto": {
            "type": "string"
          }
        }
      },
      "ProjetRequest": {
        "required": [
          "coutRequestDto",
          "descriptionRequestDto",
          "endDateRequestDto",
          "startDateRequestDto",
          "structureIdRequestDto",
          "titreRequestDto"
        ],
        "type": "object",
        "properties": {
          "idRequestDto": {
            "type": "integer",
            "format": "int32"
          },
          "referenceRequestDto": {
            "type": "string"
          },
          "titreRequestDto": {
            "maxLength": 30,
            "minLength": 2,
            "type": "string"
          },
          "descriptionRequestDto": {
            "type": "string"
          },
          "coutRequestDto": {
            "type": "number",
            "format": "double"
          },
          "startDateRequestDto": {
            "type": "string",
            "format": "date"
          },
          "endDateRequestDto": {
            "type": "string",
            "format": "date"
          },
          "structureIdRequestDto": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "FactureResponse": {
        "type": "object",
        "properties": {
          "idResponseDto": {
            "type": "integer",
            "format": "int32"
          },
          "referenceResponseDto": {
            "type": "string"
          },
          "titreResponseDto": {
            "type": "string"
          },
          "descriptionResponseDto": {
            "type": "string"
          },
          "coutResponseDto": {
            "type": "number",
            "format": "double"
          },
          "startDateResponseDto": {
            "type": "string",
            "format": "date"
          },
          "endDateResponseDto": {
            "type": "string",
            "format": "date"
          },
          "statusResponseDto": {
            "type": "string"
          },
          "termePaimentResponseDto": {
            "type": "string"
          },
          "informationFacturationResponseDto": {
            "type": "string"
          },
          "contactInformationResponseDto": {
            "type": "string"
          },
          "contratIdResponseDto": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "StructureResponse": {
        "type": "object",
        "properties": {
          "idResponseDto": {
            "type": "integer",
            "format": "int32"
          },
          "referenceResponseDto": {
            "type": "string"
          },
          "titreResponseDto": {
            "type": "string"
          },
          "descriptionResponseDto": {
            "type": "string"
          }
        }
      },
      "PageResponseProjetResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjetResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
      "ProjetResponse": {
        "type": "object",
        "properties": {
          "idResponseDto": {
            "type": "integer",
            "format": "int32"
          },
          "referenceResponseDto": {
            "type": "string"
          },
          "titreResponseDto": {
            "type": "string"
          },
          "descriptionResponseDto": {
            "type": "string"
          },
          "coutResponseDto": {
            "type": "number",
            "format": "double"
          },
          "startDateResponseDto": {
            "type": "string",
            "format": "date"
          },
          "endDateResponseDto": {
            "type": "string",
            "format": "date"
          },
          "structureIdResponseDto": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "PageResponseFactureResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FactureResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
      "ContratResponse": {
        "type": "object",
        "properties": {
          "idResponseDto": {
            "type": "integer",
            "format": "int32"
          },
          "referenceResponseDto": {
            "type": "string"
          },
          "titreResponseDto": {
            "type": "string"
          },
          "descriptionResponseDto": {
            "type": "string"
          },
          "coutResponseDto": {
            "type": "number",
            "format": "double"
          },
          "startDateResponseDto": {
            "type": "string",
            "format": "date"
          },
          "endDateResponseDto": {
            "type": "string",
            "format": "date"
          },
          "traveauxResponseDto": {
            "type": "string"
          },
          "livrablesResponseDto": {
            "type": "string"
          },
          "milestonesResponseDto": {
            "type": "string"
          },
          "termPaimentResponseDto": {
            "type": "string"
          },
          "clauseDeResiliasionResponseDto": {
            "type": "string"
          },
          "appelOffreIdResponseDto": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "PageResponseContratResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContratResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT auth description",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}